Bugfix: remove dev toggle, back/home button (#53)
