+++ app/(components)/LightDarkToggle.tsx
@@ -35,7 +35,7 @@ const LightDarkToggle = (props: LightDarkToggleProps) => {
       data-style='none'
       aria-label='toggle light/dark button'
       data-testid='light-dark-toggle-button'
-      className='opacity-100 hover:animate-pulse transition-all text-2xl lg:text-5xl dark:text-neon-blue justify-self-end'
+      className='opacity-100 hover:animate-pulse transition-all text-2xl lg:text-5xl dark:text-neon-blue'
       onClick={onToggle}
     >
       {isDark ? <MdDarkMode /> : <MdOutlineWbTwilight />}
