+++ app/head.tsx
@@ -94,13 +94,6 @@ export default function Head() {
   };
   return (
     <>
-      {/* <script
-        async
-        type='text/javascript'
-        src='https://app.termly.io/embed.min.js'
-        data-auto-block='on'
-        data-website-uuid='641b2885-7d6f-4021-9f3a-6be59c54b67f'
-      ></script> */}
       <title>{title}</title>
       <meta content='width=device-width, initial-scale=1' name='viewport' />
       <meta charSet='UTF-8' />
@@ -149,11 +142,6 @@ export default function Head() {
         name='google-site-verification'
         content='ropLCQ8cEksVS7dB6jbFu4wrAfkdeTPe05Tj2m4zZGk'
       />
-      <script
-        async
-        src='https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6537113723731394'
-        crossOrigin='anonymous'
-      ></script>
       <link rel='icon' href='/favicon.ico' />
     </>
   );
