+++ app/layout.tsx
@@ -5,13 +5,11 @@ import { Orbitron, Grenze } from '@next/font/google';
 import { AnalyticsWrapper } from './(components)/AnalayticsWrapper';
 import WordCarousell from './(components)/WordCarousell';
 import { usePathname } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import LightDarkToggle from './(components)/LightDarkToggle';
-import { getMaintenance } from '../lib/services/frontend/maintenanceService';
 import { ToastContainer } from 'react-toastify';
 import { BsFillQuestionSquareFill } from 'react-icons/bs';
 import FeaturePopup from './(components)/(FeaturePopup)/FeaturePopup';
-import LoadingMask from './(components)/LoadingMask';
 import UserDisplay from './(components)/UserDisplay';
 
 import './global.css';
@@ -19,6 +17,7 @@ import './main.css';
 import 'react-toastify/dist/ReactToastify.css';
 import 'react-confirm-alert/src/react-confirm-alert.css'; // Import css
 import Maintenance from './(components)/Maintenance';
+import BackButton from './(components)/BackButton';
 
 const grenze = Grenze({
   weight: '400',
@@ -94,7 +93,7 @@ export default function RootLayout({
       <body className='bg-black dark:bg-neon-black dark:text-neon-white text-white'>
         <header
           id='header-section'
-          className={`w-full fixed top-0 h-12 lg:h-20 gap-2 z-40 p-4 grid grid-cols-[1fr_auto_1fr] text-center items-center content-start ${
+          className={`w-full fixed top-0 h-12 lg:h-20 gap-2 z-40 p-4 flex flex-row justify-between text-center items-center ${
             pathName === '/daily-challenge' ||
             pathName === '/level' ||
             pathName === '/ai'
@@ -102,17 +101,21 @@ export default function RootLayout({
               : 'bg-black dark:bg-neon-black'
           } `}
         >
-          <Link
-            hidden={pathName !== '/'}
-            href='/rule'
-            aria-label='Link to rule page'
-            className='text-white dark:text-neon-red-light text-xl dark:text-sm lg:dark:text-2xl lg:text-3xl'
-          >
-            <div className='flex flex-row gap-2 items-center'>
-              <BsFillQuestionSquareFill data-testid='rule-icon' />
-              <span>Help</span>
-            </div>
-          </Link>
+          {pathName === '/' ? (
+            <Link
+              hidden={pathName !== '/'}
+              href='/rule'
+              aria-label='Link to rule page'
+              className='text-white dark:text-neon-red-light text-xl dark:text-sm lg:dark:text-2xl lg:text-3xl'
+            >
+              <div className='flex flex-row gap-2 items-center'>
+                <BsFillQuestionSquareFill data-testid='rule-icon' />
+                <span>Help</span>
+              </div>
+            </Link>
+          ) : (
+            <BackButton />
+          )}
           <UserDisplay />
           <LightDarkToggle
             onToggle={(dark) => {
