+++ lib/services/backend/userService.ts
@@ -8,7 +8,7 @@ import {
   selectAllUsers,
   selectUserByRecoveryKey,
 } from '../../database/userRepository';
-import { generateHashedString } from '../../utils';
+import { generateHashedString } from '../../utilities';
 import type IUser from '../../../types/user.interface';
 
 const createUser = async (
